Extract _experiment.statistics
