+++ eelbrain/_experiment/results.py
@@ -1,563 +1,24 @@
 # Author: Christian Brodbeck <christianbrodbeck@nyu.edu>
-"""Result and movie derivatives.
+"""Movie export derivatives.
 
 These derivatives orchestrate through other graph nodes, especially the
 dataset-producing derivatives that correspond to the public ``Pipeline.load_x``
-methods. Cache paths are internal derivative-owned artifacts; only end-product
-exports such as movies own default public paths. They must not depend on
-injected facade loaders or Pipeline-managed naming state, and naming-only path
-labels must not be part of canonical cache identity.
+methods. They share the result-output machinery in
+:mod:`._experiment.statistics.nodes`; only end-product exports such as movies
+own default public paths.
 """
 
 from __future__ import annotations
 
-import logging
 from pathlib import Path
-from typing import Any, TypeVar
+from typing import Any
 
-from .. import load
 from .. import plot
-from .. import save
 from .. import testnd
-from .._data_obj import Dataset
-from .._exceptions import ConfigurationError
-from .._io.pickle import update_subjects_dir
-from .._text import enumeration
 from .._stats.stats import ttest_t
-from .._stats.testnd import _MergedTemporalClusterDist
-from .derivative_cache import Dependency, Derivative, Request, UncachedDerivative
-from .pathing import (
-    MRI_SDIR,
-    epoch_basename,
-    join_stem_parts,
-    movie_export_path,
-    report_export_path,
-    test_basename,
-    time_window_str,
-)
-from .source import ROIData, roi_data_from_subject_datasets
-from .test_def import ROITestResult, ResolvedTestNDSpec, Test, DataSpec
-from .variable_def import apply_vardef
-
-T = TypeVar('T')
-USE_CTX = object()
-RESULT_OPTION_DEFAULTS = {
-    'samples': None,
-    'data': None,
-    'test': None,
-    'tstart': None,
-    'tstop': None,
-    'pmin': None,
-    'baseline': None,
-    'src_baseline': None,
-    'smooth': None,
-    'samplingrate': None,
-}
-
-TEST_DATA_OPTION_NAMES = (
-    'data',
-    'test',
-    'baseline',
-    'src_baseline',
-    'samplingrate',
-    'smooth',
-)
-
-
-def _test_result_options(
-        ctx: Request,
-        *,
-        data: DataSpec | object = USE_CTX,
-) -> dict[str, Any]:
-    if data is USE_CTX:
-        data = ctx.options['data']
-    out = {
-        'data': data,
-        'samples': ctx.options['samples'],
-        'test': ctx.options['test'],
-        'tstart': ctx.options['tstart'],
-        'tstop': ctx.options['tstop'],
-        'pmin': ctx.options['pmin'],
-        'baseline': ctx.options['baseline'],
-        'src_baseline': ctx.options['src_baseline'],
-        'smooth': ctx.options['smooth'],
-        'samplingrate': ctx.options['samplingrate'],
-    }
-    if 'disconnect_labels' in ctx.options:
-        out['disconnect_labels'] = ctx.options['disconnect_labels']
-    return out
-
-
-def _evoked_stc_options(
-        ctx: Request,
-        baseline=USE_CTX,
-        src_baseline=USE_CTX,
-        morph: bool = False,
-        cat=None,
-        samplingrate: int | None = None,
-        decim: int | None = None,
-        ndvar: bool = True,
-) -> dict[str, Any]:
-    if baseline is USE_CTX:
-        baseline = ctx.options['baseline']
-    if src_baseline is USE_CTX:
-        src_baseline = ctx.options['src_baseline']
-    return ctx.options_for(
-        'evoked-stc',
-        baseline=baseline,
-        src_baseline=src_baseline,
-        morph=morph,
-        cat=cat,
-        samplingrate=samplingrate,
-        decim=decim,
-        ndvar=ndvar,
-        keep_evoked=False,
-    )
-
-
-def _epochs_stc_options(
-        ctx: Request,
-        baseline=USE_CTX,
-        src_baseline=USE_CTX,
-        cat=None,
-        keep_epochs: bool | str = False,
-        morph: bool | None = None,
-        samplingrate: int | None = None,
-        decim: int | None = None,
-        ndvar: bool = True,
-        reject: bool | str = True,
-) -> dict[str, Any]:
-    if baseline is USE_CTX:
-        baseline = ctx.options['baseline']
-    if src_baseline is USE_CTX:
-        src_baseline = ctx.options['src_baseline']
-    return ctx.options_for(
-        'epochs-stc',
-        baseline=baseline,
-        src_baseline=src_baseline,
-        cat=cat,
-        keep_epochs=keep_epochs,
-        morph=morph,
-        samplingrate=samplingrate,
-        decim=decim,
-        ndvar=ndvar,
-        reject=reject,
-    )
-
-
-def _validate_post_aggregation_test_vars(test_obj: Test, data_desc: str):
-    model_vars = set(filter(None, (test_obj.model or '').split('%')))
-    missing_model_vars = sorted(model_vars.intersection(test_obj.vars.vars))
-    if missing_model_vars:
-        vars_desc = enumeration(missing_model_vars)
-        raise ConfigurationError(
-            f"For evoked-backed {data_desc} tests, Test.vars must be computable from the post-aggregation dataset. "
-            f"Model variable {vars_desc} can not be provided through Test.vars. Use TwoStageTest or Pipeline.variables instead."
-        )
-
-
-def _apply_post_aggregation_test_vars(ds, test_obj: Test, tests, groups, data_desc: str):
-    if not test_obj.vars:
-        return ds
-    _validate_post_aggregation_test_vars(test_obj, data_desc)
-    try:
-        apply_vardef(ds, test_obj.vars, tests, groups)
-    except Exception as error:
-        raise ConfigurationError(f"For evoked-backed {data_desc} tests, Test.vars must be computable from the post-aggregation dataset. Use TwoStageTest or Pipeline.variables for trial-level variables ({error}).") from None
-    return ds
-
-
-class ResultOutputDerivative(Derivative[T]):
-    """Shared base for cached result/report/movie outputs.
-
-    This is a :class:`~eelbrain._experiment.derivative_cache.Derivative`
-    subclass with a fixed pattern:
-
-    - :meth:`key` encodes the logical analysis identity, independent of any
-      explicit output destination.
-    - :meth:`fingerprint` delegates to :meth:`Derivative.standard_fingerprint`
-      using shared result-state and configured test/epoch/parc definitions.
-    - :meth:`path` chooses a user-facing export path, with optional
-      ``samples``-specific disambiguation.
-    - :meth:`load` returns that path, and :meth:`save` is a no-op, because
-      subclasses normally create the final output file directly in
-      :meth:`build` rather than serializing a separate in-memory artifact.
-
-    The underscored helper methods are grouped by the derivative hook they
-    support:
-
-    - ``_key_*`` helpers feed :meth:`key`
-    - ``_fingerprint_*`` helpers feed :meth:`fingerprint`
-    - ``_path_*`` helpers feed :meth:`path`
-
-    Subclasses usually extend this template by overriding:
-
-    - :meth:`dependencies` and :meth:`build` as ordinary derivative hooks
-    - :meth:`_identity_extra` to add result-family-specific identity fields
-    - :meth:`_path_stem` or :meth:`_default_output_path` to customize export
-      naming
-
-    Options
-    -------
-    dst
-        Optional explicit output path.
-    samples
-        Permutation/sample count stored in the cache identity.
-    data
-        Analysis data family to use (sensor, source, ROI, ...).
-    test
-        Test definition to run.
-    tstart, tstop
-        Optional time window for the analysis.
-    pmin
-        Cluster-forming threshold or ``'tfce'``.
-    baseline
-        Sensor-space baseline correction.
-    src_baseline
-        Source-space baseline correction.
-    disconnect_labels
-        Disconnect source-space cluster adjacency across labels from the
-        current ``parc`` state.
-    samplingrate
-        Sampling rate override for upstream cached data.
-    smooth
-        Optional source-space smoothing.
-    """
-    key_fields = ()
-    cache_log_level = logging.INFO
-    single_subject = False
-    sampled_path = False
-    OPTION_DEFAULTS = RESULT_OPTION_DEFAULTS
-    VIEW_OPTION_DEFAULTS = {'dst': None}
-
-    def __init__(
-            self,
-            tests: dict[str, Test],
-            epochs: dict[str, Any],
-            parcs: dict[str, Any],
-            groups: dict[str, tuple[str, ...] | list[str]],
-    ):
-        self.tests = tests
-        self.epochs = epochs
-        self.parcs = parcs
-        self.groups = groups
-
-    def _key_state_snapshot(
-            self,
-            ctx: Request,
-            single_subject: bool,
-    ) -> dict[str, Any]:
-        """Canonical state subset used by :meth:`key`."""
-        data = ctx.options['data']
-        fields = ['epoch', 'raw', 'epoch_rejection', 'model', 'equalize_evoked_count', 'test']
-        if data and data.source:
-            fields.extend(['cov', 'inv', 'src', 'mri', 'parc'])
-        else:
-            fields.append('reference')
-        state = {field: ctx.state[field] for field in fields}
-        if single_subject:
-            state['subject'] = ctx.state['subject']
-        else:
-            state['subjects'] = tuple(self.groups[ctx.state['group']])
-        return ctx.registry.canonicalize(state)
-
-    def _key_analysis_options(self, ctx: Request) -> dict[str, Any]:
-        """Canonical analysis options used by :meth:`key`."""
-        data = ctx.options['data']
-        return ctx.registry.canonicalize({
-            'data': None if data is None else data.string,
-            'samples': ctx.options['samples'],
-            'baseline': ctx.options['baseline'],
-            'src_baseline': ctx.options['src_baseline'],
-            'disconnect_labels': ctx.options.get('disconnect_labels', False),
-            'pmin': ctx.options['pmin'],
-            'tstart': ctx.options['tstart'],
-            'tstop': ctx.options['tstop'],
-            'samplingrate': ctx.options['samplingrate'],
-            'smooth': ctx.options['smooth'],
-            'adjacency': ctx.state['adjacency'],
-        })
-
-    def _key_identity(self, ctx: Request) -> dict[str, Any]:
-        """Stable logical identity shared by result-output cache keys."""
-        return ctx.registry.canonicalize({
-            'state': self._key_state_snapshot(ctx, self.single_subject),
-            'options': self._key_analysis_options(ctx),
-            'single_subject': self.single_subject,
-            **self._identity_extra(ctx),
-        })
-
-    def _path_context_parts(self, ctx: Request) -> list[str]:
-        """Path-stem parts derived from analysis context/state."""
-        data = ctx.options['data']
-        parts = [f'data-{data.string}', f'raw-{ctx.state["raw"]}', f'rej-{ctx.state["epoch_rejection"]}']
-        if ctx.state['model']:
-            parts.append(f'model-{ctx.state["model"]}')
-        if ctx.state['equalize_evoked_count']:
-            parts.append(f'count-{ctx.state["equalize_evoked_count"]}')
-        if data.source:
-            parts.extend((f'cov-{ctx.state["cov"]}', f'src-{ctx.state["src"]}', f'inv-{ctx.state["inv"]}', f'parc-{ctx.state["parc"]}'))
-        return parts
-
-    def _path_option_parts(self, ctx: Request) -> list[str]:
-        """Path-stem parts derived from analysis options."""
-        parts = []
-        baseline = ctx.options['baseline']
-        src_baseline = ctx.options['src_baseline']
-        pmin = ctx.options['pmin']
-        samplingrate = ctx.options['samplingrate']
-        smooth = ctx.options['smooth']
-        if baseline is False:
-            parts.append('nobl')
-        elif baseline not in (None, True):
-            parts.append(f'bl-{time_window_str(baseline)}')
-        if src_baseline is True:
-            parts.append('srcbl')
-        elif src_baseline not in (None, False):
-            parts.append(f'srcbl-{time_window_str(src_baseline)}')
-        if ctx.options.get('disconnect_labels', False):
-            parts.append('disconnect-labels')
-        if pmin == 'tfce':
-            parts.append('tfce')
-        elif pmin is not None:
-            parts.append(f'p-{pmin}')
-        if pmin is not None and ctx.options['data'].source and ctx.state['adjacency']:
-            parts.append(f'adj-{ctx.state["adjacency"]}')
-        if ctx.options['tstart'] is not None or ctx.options['tstop'] is not None:
-            parts.append(f'tw-{time_window_str((ctx.options["tstart"], ctx.options["tstop"]))}')
-        if samplingrate is not None:
-            parts.append(f'sr-{samplingrate:g}Hz')
-        if smooth:
-            parts.append(f'sm-{int(round(smooth * 1000))}mm')
-        return parts
-
-    def _path_stem(self, ctx: Request) -> str:
-        """Default export stem used by :meth:`path`."""
-        return join_stem_parts(
-            test_basename(ctx.state),
-            f'epoch-{ctx.state["epoch"]}',
-            f'test-{ctx.options["test"]}',
-            self._path_context_parts(ctx),
-            self._path_option_parts(ctx),
-        )
-
-    def _default_output_path(self, ctx: Request) -> Path:
-        """Default user-facing export path used when ``dst`` is not set."""
-        return ctx.root / report_export_path(ctx.state, self.name, self._path_stem(ctx), self.single_subject)
-
-    def _identity_extra(self, ctx: Request) -> dict[str, Any]:
-        """Extra identity fields shared by :meth:`key` and :meth:`fingerprint`."""
-        return {}
-
-    def key(self, ctx: Request) -> dict[str, Any]:
-        return {
-            'identity': self._key_identity(ctx),
-            'options': ctx.options,
-        }
-
-    def fingerprint(self, ctx: Request) -> dict[str, Any]:
-        out = {
-            'test': self.tests[ctx.options['test']],
-            'epoch': self.epochs[ctx.state['epoch']],
-            'single_subject': self.single_subject,
-            **self._identity_extra(ctx),
-        }
-        if not self.single_subject:
-            out['subjects'] = self.groups[ctx.state['group']]
-        if ctx.options['data'].source and ctx.state['parc'] in self.parcs:
-            out['parc'] = self.parcs[ctx.state['parc']]
-        return out
-
-    def path(
-            self,
-            ctx: Request,
-    ) -> Path:
-        dst = ctx.view_options['dst']
-        path = Path(dst) if dst else self._default_output_path(ctx)
-        if self.sampled_path:
-            path = sampled_artifact_path(path, ctx.options['samples'])
-        return path
-
-    def load(
-            self,
-            ctx: Request,
-            path: Path) -> T:
-        return path
-
-    def save(
-            self,
-            ctx: Request,
-            path: Path,
-            value: T,
-    ) -> None:
-        return
-
-
-def sampled_artifact_path(path: str | Path, samples: int | None) -> Path:
-    path = Path(path)
-    if samples is None:
-        return path
-    return path.with_name(f"{path.stem}_samples-{samples}{path.suffix}")
-
-
-class EvokedTestDataDerivative(UncachedDerivative[Dataset | ROIData]):
-    """Prepared test/report data for sensor, source, and ROI analyses.
-
-    Options
-    -------
-    data
-        Analysis data family to prepare.
-    baseline
-        Sensor-space baseline correction.
-    src_baseline
-        Source-space baseline correction.
-    samplingrate
-        Sampling rate override for upstream cached data.
-    smooth
-        Optional source-space smoothing.
-    """
-    name = 'evoked-test-data'
-    OPTION_DEFAULTS = {
-        'data': None,
-        'test': None,
-        'baseline': None,
-        'src_baseline': None,
-        'samplingrate': None,
-        'smooth': None,
-    }
-
-    def __init__(self, tests: dict[str, Test], epochs: dict[str, Any], groups: dict[str, tuple[str, ...] | list[str]]):
-        self.tests = tests
-        self.epochs = epochs
-        self.groups = groups
-
-    def _sensor_evoked_options(self, ctx: Request, cat) -> dict[str, Any]:
-        return ctx.options_for(
-            'evoked',
-            baseline=ctx.options['baseline'],
-            ndvar=True,
-            cat=cat,
-            samplingrate=ctx.options['samplingrate'],
-            decim=None,
-            data=ctx.options['data'],
-        )
-
-    def fingerprint(self, ctx: Request) -> dict[str, Any]:
-        return {
-            'test': self.tests[ctx.options['test']],
-            'epoch': self.epochs[ctx.state['epoch']],
-            'subjects': tuple(self.groups[ctx.state['group']]),
-        }
-
-    def dependencies(self, ctx: Request) -> tuple[Dependency, ...]:
-        data = ctx.options['data']
-        test_obj = self.tests[ctx.options['test']]
-        samplingrate = ctx.options['samplingrate']
-        subjects = self.groups[ctx.state['group']]
-
-        if data.sensor:
-            return (Dependency('evoked-group-dataset', options=self._sensor_evoked_options(ctx, test_obj.cat)),)
-
-        if data.source and not data.aggregate:
-            return (Dependency(
-                'evoked-stc-group-dataset',
-                options=_evoked_stc_options(ctx, morph=True, cat=test_obj.cat, samplingrate=samplingrate),
-            ),)
-
-        return tuple(
-            Dependency(
-                'evoked-stc',
-                label=subject,
-                state={'subject': subject},
-                options=_evoked_stc_options(ctx, morph=False, cat=None, samplingrate=samplingrate),
-            )
-            for subject in subjects
-        )
-
-    def build(self, ctx: Request) -> Dataset | ROIData:
-        data = ctx.options['data']
-        test_obj = self.tests[ctx.options['test']]
-        subjects = self.groups[ctx.state['group']]
-        if test_obj.vars:
-            _validate_post_aggregation_test_vars(test_obj, data.string)
-
-        if data.sensor:
-            ds = ctx.load('evoked-group-dataset')
-            return _apply_post_aggregation_test_vars(ds, test_obj, self.tests, self.groups, data.string)
-
-        if data.source and not data.aggregate:
-            ds = ctx.load('evoked-stc-group-dataset')
-            ds = _apply_post_aggregation_test_vars(ds, test_obj, self.tests, self.groups, data.string)
-            if smooth := ctx.options['smooth']:
-                ds[data.y_name] = ds[data.y_name].smooth('source', smooth, 'gaussian')
-            return ds
-
-        if ctx.options['smooth']:
-            raise TypeError(f"smooth={ctx.options['smooth']!r} for ROI tests")
-
-        dss = []
-        for subject in subjects:
-            ds = ctx.load(subject)
-            dss.append(_apply_post_aggregation_test_vars(ds, test_obj, self.tests, self.groups, data.string))
-        return roi_data_from_subject_datasets(dss, data.source)
-
-
-class TestResultDerivative(ResultOutputDerivative):
-    """Cached statistical test result.
-
-    Uses the shared result-output options from
-    :class:`ResultOutputDerivative`.
-    """
-    name = 'test-result'
-    sampled_path = True
-    cache_suffix = '.pickle'
-    path = Derivative.path
-    OPTION_DEFAULTS = {**RESULT_OPTION_DEFAULTS, 'disconnect_labels': False}
-    VIEW_OPTION_DEFAULTS = {}
-
-    def cache_label(self, ctx: Request) -> str:
-        return join_stem_parts(self._path_stem(ctx), f'samples-{ctx.options["samples"]}') if ctx.options['samples'] is not None else self._path_stem(ctx)
-
-    def dependencies(self, ctx: Request) -> tuple[Dependency, ...]:
-        return (Dependency('evoked-test-data', options=ctx.options_for('evoked-test-data', *TEST_DATA_OPTION_NAMES)),)
-
-    def build(self, ctx: Request):
-        test_obj = self.tests[ctx.options['test']]
-        data = ctx.options['data']
-        test_spec = ResolvedTestNDSpec.from_request(ctx, data)
-        data_value = ctx.load('evoked-test-data')
-        if isinstance(data_value, ROIData):
-            subjects = list(self.groups[ctx.state['group']])
-            n_per_label = {label: len(ds['subject'].cells) for label, ds in data_value.label_data.items()}
-            do_mcc = len(data_value.label_data) > 1 and ctx.options['pmin'] not in (None, 'tfce') and len(set(n_per_label.values())) == 1
-            label_results = {
-                label: test_spec.make_result(self, 'label_tc', ds, test_obj, do_mcc)
-                for label, ds in data_value.label_data.items()
-            }
-            merged_dist = _MergedTemporalClusterDist([res._cdist for res in label_results.values()]) if do_mcc else None
-            return ROITestResult(subjects, ctx.options['samples'], data_value.n_trials_ds, merged_dist, label_results)
-        if data.sensor and len(data_value.info['sensor_types']) > 1:
-            desc = ', '.join(data_value.info['sensor_types'])
-            raise RuntimeError(f"Data contains more than one sensor type ({desc}). Mass-univariate tests are not designed for multiple sensor types. Use the data argument to perform test on one sensor type.")
-        return test_spec.make_result(self, test_spec.data.y_name, data_value, test_obj)
-
-    def load(
-            self,
-            ctx: Request,
-            path: Path):
-        res = load.unpickle(path)
-        if ctx.options['data'].source and not ctx.options['data'].aggregate:
-            update_subjects_dir(res, ctx.root / MRI_SDIR, 2)
-        return res
-
-    def save(
-            self,
-            ctx: Request,
-            path: Path,
-            value,
-    ) -> None:
-        save.pickle(value, path)
-
+from .derivative_cache import Dependency, Request
+from .pathing import epoch_basename, join_stem_parts, movie_export_path
+from .statistics.nodes import RESULT_OPTION_DEFAULTS, ResultOutputDerivative, _epochs_stc_options, _evoked_stc_options
 
 _MOVIE_OPTION_DEFAULTS = {
     'time_dilation': 1.0,
