+++ eelbrain/_experiment/statistics/config.py
@@ -1,23 +1,23 @@
 # Author: Christian Brodbeck <christianbrodbeck@nyu.edu>
+"""Statistical test definitions (:class:`Configuration` classes)."""
 from __future__ import annotations
 
 from dataclasses import dataclass
-from inspect import getfullargspec
 import re
 from collections.abc import Collection
 from typing import TYPE_CHECKING, Any
 
-import mne
-
-from .. import testnd
-from .. import test
-from .._data_obj import CellArg, Dataset, NDVar, Var, combine
-from .._exceptions import ConfigurationError
-from .configuration import Configuration
-from .variable_def import Variables, VarDef, GroupVar
+from ... import testnd
+from ... import test
+from ..._data_obj import CellArg, Dataset, NDVar, Var, combine
+from ..._exceptions import ConfigurationError
+from ..._utils.parse import find_variables
+from ..configuration import Configuration
+from ..data import DataSpec
+from ..variable_def import Variables, VarDef, GroupVar
 
 if TYPE_CHECKING:
-    from .derivative_cache import Request
+    from ..derivative_cache import Request
 
 
 __test__ = False
@@ -363,106 +363,83 @@ class ANOVA(Test):
         return test.ANOVA(y, self.x, data=ds)
 
 
-class DataSpec:
-    """Internal description of the data going into an analysis
+class TwoStageTest(Test):
+    """Two-stage test: T-test of regression coefficients
 
-    Combines the data *space* (sensor vs source, determined by the ``inv``
-    state) with the data *kind* (sensor type and/or aggregation). Instances are
-    composed at the :class:`Pipeline` boundary from the user-facing ``data``
-    argument and the ``inv`` state; see ``Pipeline._resolve_data``.
+    Stage 1: fit a regression model to the data for each subject.
+    Stage 2: test coefficients from stage 1 against 0 across subjects.
 
     Parameters
     ----------
-    string : str
-        Internal string describing the data: ``'sensor'``/``'source'`` (or a
-        specific sensor type ``'meg'``/``'mag'``/``'grad'``/``'eeg'``), with an
-        optional ``.mean``/``.rms`` aggregation suffix.
-    time : bool
-        Whether the base data contains a time axis.
-    morph : bool
-        If loading source space data, whether the data is morphed to the common
-        brain.
+    stage_1 : str
+        Stage 1 model specification. Coding for categorial predictors uses 0/1 dummy
+        coding.
+    vars : dict
+        Add new variables for the stage 1 model. This is useful for specifying
+        coding schemes based on categorial variables.
+        Each entry specifies a variable with the following schema:
+        ``{name: definition}``. ``definition`` can be either a string that is
+        evaluated in the events-:class:`Dataset`, or a
+        ``(source_name, {value: code})``-tuple (see example below).
+        ``source_name`` can also be an interaction, in which case cells are joined
+        with spaces (``"f1_cell f2_cell"``).
+    model : str
+        This parameter can be supplied to perform stage 1 tests on condition
+        averages. If ``model`` is not specified, the stage1 model is fit on single
+        trial data.
+
+    See Also
+    --------
+    Pipeline.tests
+
+    Examples
+    --------
+    The first example assumes 2 categorical variables present in events,
+    'a' with values 'a1' and 'a2', and 'b' with values 'b1' and 'b2'. These are
+    recoded into 0/1 codes::
+
+        TwoStageTest(
+            "a_num + b_num + a_num * b_num + index + a_num * index",
+            vars={
+                'a_num': ('a', {'a1': 0, 'a2': 1}),
+                'b_num': ('b', {'b1': 0, 'b2': 1}),
+            }),
+
+    The second test definition uses the "index" variable which is always present
+    and specifies the chronological index of the events as an integer count.
+    This variable can thus be used to test for a linear change over time. Due
+    to the numeric nature of these variables interactions can be computed by
+    multiplication::
+
+        TwoStageTest("a_num + index + a_num * index",
+                     vars={'a_num': ('a', {'a1': 0, 'a2': 1})
+
+    Numerical variables can also defined using data-object methods (e.g.
+    :meth:`Factor.label_length`) or from interactions::
+
+        TwoStageTest('wordlength', vars={'wordlength': 'word.label_length()'})
+        TwoStageTest("ab", vars={'ab': ('a%b', {'a1 b1': 0, 'a1 b2': 1, 'a2 b1': 1, 'a2 b2': 2})})
     """
-    RE = re.compile(r"^(source|sensor|meg|mag|grad|eeg)(?:\.(mean|rms))?$")
-    source = False
-    sensor = False
-    aggregate = None  # None, 'mean', or 'rms'
+    kind = 'two-stage'
+    DICT_ATTRS = Test.DICT_ATTRS + ('stage_1',)
 
-    def __init__(self, string, time=True, morph=False):
-        self.time = bool(time)
-        self.morph = bool(morph)
-        m = self.RE.match(string)
-        if m is None:
-            raise ValueError(f"data={string!r}: invalid data description")
-        dim, self.aggregate = m.groups()
-        if dim in ('meg', 'mag'):
-            self._to_ndvar = ('mag',)
-            self.y_name = 'meg'  # mag NDVars are keyed 'meg' (see .load_epochs())
-            self.sensor = True
-        elif dim in ('grad', 'eeg'):
-            self._to_ndvar = (dim,)
-            self.y_name = dim
-            self.sensor = True
-        elif dim == 'sensor':
-            self._to_ndvar = None
-            self.y_name = 'meg'
-            self.sensor = True
-        elif dim == 'source':
-            self._to_ndvar = None
-            self.y_name = 'srcm' if self.morph else 'src'
-            self.source = True
-        else:
-            raise RuntimeError(f"{string=} ({dim=})")
-        self.string = string
+    def __init__(self, stage_1: str, vars: dict = None, model: str = None):
+        Test.__init__(self, stage_1, model, vars=vars, depend_on=find_variables(stage_1))
+        self.stage_1 = stage_1
 
-        dims = []
-        if self.source and not self.aggregate:
-            dims.append('source')
-        elif self.sensor and not self.aggregate:
-            dims.append('sensor')
-        if self.time:
-            dims.append('time')
-        self.dims = tuple(dims)
+    def make_stage_1(self, y, data, subject, sub=None):
+        """Assumes that model has already been applied"""
+        return testnd.LM(y, self.stage_1, sub=sub, data=data, samples=0, subject=subject)
 
-        # whether parc is used from subjects or from common-brain
-        if self.source and not self.aggregate:
-            self.parc_level = 'common'
-        elif self.source:
-            self.parc_level = 'individual'
-        else:
-            self.parc_level = None
+    @staticmethod
+    def make_stage_2(lms, kwargs):
+        lm = testnd.LMGroup(lms)
+        lm.compute_column_ttests(**kwargs)
+        return lm
 
-    @classmethod
-    def coerce(cls, obj, time=True, morph=False):
-        if isinstance(obj, cls):
-            if obj.time == time and obj.morph == morph:
-                return obj
-            else:
-                return cls(obj.string, time, morph)
-        else:
-            return cls(obj, time, morph)
-
-    def __repr__(self):
-        return f"DataSpec({self.string!r})"
-
-    def __eq__(self, other):
-        if not isinstance(other, DataSpec):
-            return False
-        return self.string == other.string and self.time == other.time
-
-    def _testnd_parc(self, disconnect_labels: bool) -> str | None:
-        if self.source and not self.aggregate:
-            return 'source' if disconnect_labels else None
-        if disconnect_labels:
-            raise TypeError(f"{disconnect_labels=}: invalid for data={self.string!r}")
-        return None
-
-    def data_to_ndvar(self, info: mne.Info) -> list[str]:
-        assert self.sensor
-        if self._to_ndvar is None:
-            return info.get_channel_types(unique=True, only_data_chs=True)
-        else:
-            return self._to_ndvar
+    def make(self, y, ds, force_permutation, kwargs):
+        lms = [self.make_stage_1(y, ds, subject, f"subject=={subject!r}") for subject in ds['subject'].cells]
+        return self.make_stage_2(lms, kwargs)
 
 
 @dataclass(frozen=True)
@@ -514,33 +491,3 @@ class ResolvedTestNDSpec:
         if isinstance(y, NDVar) and y.has_dim('space'):
             return test_obj._make_vec(y, ds, force_permutation, self.kwargs)
         return test_obj._make(y, ds, force_permutation, self.kwargs)
-
-
-class ROITestResult:
-    """Test results for temporal tests in one or more ROIs
-
-    Attributes
-    ----------
-    subjects : tuple of str
-        Subjects included in the test.
-    samples : int
-        ``samples`` parameter used for permutation tests.
-    res : {str: NDTest} dict
-        Test result for each ROI.
-    n_trials_ds : Dataset
-        Dataset describing how many trials were used in each condition per
-        subject.
-    """
-
-    def __init__(self, subjects, samples, n_trials_ds, merged_dist, res):
-        self.subjects = subjects
-        self.samples = samples
-        self.n_trials_ds = n_trials_ds
-        self.merged_dist = merged_dist
-        self.res = res
-
-    def __getstate__(self):
-        return {attr: getattr(self, attr) for attr in getfullargspec(self.__init__).args[1:]}
-
-    def __setstate__(self, state):
-        self.__init__(**state)
